+++ CHANGES.md
@@ -1,5 +1,12 @@
 # Changes
 
+## 0.0.6 (2026-04-05)
+
+- Added TOML support
+- Fixed error reporting
+- Added no sandbox flag
+- Added package exclusion
+
 ## 0.0.5 (2026-04-05)
 
 - Added Docker support